[ONNX] Update LayerNorm symbolic to handle autocasting (#52199)
When onnx export creates a 0-dim tensor of constant type, this action overrides the type promotion logic as quoted in #9515. In order to prevent this from happening this PR adds the following functionality.
If the data type is a floating point type, it is converted to a 0-dim double tensor, else it is converted to a 0-dim tensor of its original type

ghstack-source-id: 189ee88787f6a70618efd1df20b4903dd4c14cf8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52350